# lib/api-zod/src/generated/types/practiceAssignmentGrade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeAssignmentResult } from './practiceAssignmentResult';

export interface PracticeAssignmentGrade {
  score: number;
  total: number;
  percent: number;
  results: PracticeAssignmentResult[];
}