# lib/api-zod/src/generated/types/practiceAssignmentProblem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeAssignmentProblem {
  id: number;
  prompt: string;
  topicId: number;
  /** @nullable */
  topicTitle?: string | null;
}