# lib/api-zod/src/generated/types/practiceAssignmentResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeAssignmentResult {
  problemId: number;
  prompt: string;
  correct: boolean;
  userAnswer?: string;
  correctAnswer: string;
  feedback: string;
}