# lib/api-zod/src/generated/types/practiceAssignmentSet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeAssignmentProblem } from './practiceAssignmentProblem';

export interface PracticeAssignmentSet {
  sessionId: number;
  assignmentId: number;
  title: string;
  kind: string;
  /** @nullable */
  instructions: string | null;
  problems: PracticeAssignmentProblem[];
}